renamings and better setup
