+++ setup.py
@@ -7,5 +7,8 @@ setuptools.setup(
     author_email='andrey.naradzetski@gmail.com',
     url='https://github.com/anaradzetski/telegram_recursive_keyboard',
     packages=['telegram_utils'],
+    install_requires=[
+        'python-telegram-bot>=13.5'
+    ],
     python_requires='>=3.6'
 )
